fix(bb-cron-job): validate the schedule/timezone at synth, not at deploy
The CDK layer passed schedule/timezone straight into the EventBridge CfnSchedule
with no validation, so an invalid expression (e.g. rate(10 seconds), whose
minimum is 1 minute, or a malformed cron) passed cdk synth and was only rejected
by EventBridge minutes into provisioning — while the mock already validated it,
so local dev and deploy diverged.

Extract the rate/cron parser + timezone check into a shared schedule module used
by both the mock and the CDK construct. CronJob's constructor now validates up
front and throws CronJobErrors.InvalidSchedule / InvalidTimezone at synth, before
any infrastructure is created.
